requirements.txt
logzero
python-dotenv
psycopg2-binary == 2.9.5
SQLAlchemy == 1.4.46
git+https://github.com/agronholm/sqlacodegen.git
click == 8.1.3

celery == 5.2.7
celery-singleton == 0.3.1
flower == 1.2.0
billiard ==  3.6.4.0 
redis

arrow == 1.2.3

lxml == 4.9.2
requests == 2.28.2
beautifulsoup4 == 4.11.1
praw == 7.6.1

retry == 0.9.2
tqdm == 4.64.1
setuptools == 66.1.1

numpy == 1.24.1
pandas == 1.5.3
matplotlib == 3.6.3
Pillow == 9.4.0
opencv-python-headless == 4.7.0.68

types-tqdm == 4.64.7.11